Move LineItems outside of Import namespace
